Added integration with backend
