+++ src/constants/material-type.js
@@ -5,7 +5,7 @@ export const MATERIAL_TYPE_PATH = `/${SETTINGS_PATH}/${MATERIAL_TYPE_ROUTE}`;
 
 export const MATERIAL_TYPE_FIELDS = {
   ID: 'id',
-  MATERIAL_TYPE_MAPPING_LIST: 'materialTypeMapping',
+  MATERIAL_TYPE_MAPPING_LIST: 'materialTypeMappings',
   CENTRAL_SERVER_ID: 'centralServerId',
   CENTRAL_ITEM_TYPE: 'centralItemType',
   MATERIAL_TYPE_ID: 'materialTypeId',
